Prevent OOM in read_from_client() for large uploads
CGI.pm calls ->read_from_client() with the whole Content-Length.
In CGI.pm (or under mod_perl) it does not matter, as the socket
filehandle always returns partial read, and thus the chunk read
into memory is not excessively large.

On the other hand, psgi.input handle can be buffered, and tries to
satisfy the whole ->read() at once. This can lead to excessive memory
usage even in situations, which CGI.pm tries to prevent (reading uploads
to temporary files instead of memory, etc.).

The fix is to impose the upper limit of bytes read in one call
to read_from_client().